Refactored FillRect artist into its own package to make room for new artists.
